Add little bit of documentation
The README has been updated to contain a little bit more information.

A markdown file for the final report has been added. To simplify the
generation of a PDF file there is also a Makefile now. To build the PDF
you will need a working pandoc installation and the eisvogel LaTeX
template [1].

The gitignore has been adjusted to prevend PDFs from beeing committed.

[1]: https://github.com/Skyfmmf/pandoc-latex-template